More refactoring. Now data holders are in the same class. Also making controller inject a GooglePLacesObject, what will make it much easier to test controller
